requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.8.2
Pillow==10.4.0
aiofiles==23.2.1
